Mystery Calculator
Ask someone to choose a number from any of the six cards. Show them each card in turn and ask them if their number appears on it. Print out the list of numbers in the card. Add the numbers in the top left-hand corner of each card that contains their number. The total is their number. Output the user's number